Update on "move get tokenizer to export_llama_lib"
Differential Revision: [D60419366](https://our.internmc.facebook.com/intern/diff/D60419366/)

[ghstack-poisoned]
